Bug fixes
Fixed a rounding error with time calculations that was compounded by
time acceleration.

Fixed a major bug which threw index out of bounds errors for puzzles
supporting more than 4 teams.

+++ Assets/Scripts/Player.cs
@@ -59,7 +59,7 @@ public class Player : MonoBehaviour {
 		float horiz = 0;
 		float vert = 0;
 
-		float velocity = moveSpeed * GameController.getDeltaTime ();
+		float velocity = moveSpeed * (float) GameController.getDeltaTime ();
 
 		if (!transform.Equals (moveTarget)) {
 			// move to target
